Refactor the AI Popover into a new component
