+++ chainforge/react-server/src/TextFieldsNode.js
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useEffect, useCallback } from 'react';
 import { Handle } from 'reactflow';
-import { Textarea, Tooltip, TextInput, NumberInput, Button, Text, Stack, Tabs, Switch, Skeleton } from '@mantine/core';
+import { Textarea, Tooltip, Skeleton } from '@mantine/core';
 import { IconTextPlus, IconEye, IconEyeOff } from '@tabler/icons-react';
 import useStore from './store';
 import NodeLabel from './NodeLabelComponent';
@@ -9,6 +9,7 @@ import BaseNode from './BaseNode';
 import { setsAreEqual } from './backend/utils';
 import { AIError, autofill, generateAndReplace } from './backend/ai';
 import { debounce } from 'lodash';
+import AiPopover from './AiPopover';
 
 // Helper funcs
 const union = (setA, setB) => {
@@ -31,19 +32,11 @@ const TextFieldsNode = ({ data, id }) => {
   const setDataPropsForNode = useStore((state) => state.setDataPropsForNode);
   const pingOutputNodes = useStore((state) => state.pingOutputNodes);
 
-  const [commandFillNumber, setCommandFillNumber] = useState(3);
-
-  const [generateAndReplaceNumber, setGenerateAndReplaceNumber] = useState(3);
-  const [generateAndReplacePrompt, setGenerateAndReplacePrompt] = useState('');
-  const [generateAndReplaceIsUnconventional, setGenerateAndReplaceIsUnconventional] = useState(false);
-  const [generateAndReplaceIsLoading, setGenerateAndReplaceIsLoading] = useState(false);
-  const [didGenerateAndReplaceError, setDidGenerateAndReplaceError] = useState(false);
-
-  const [isCommandFillLoading, setIsCommandFillLoading] = useState(false);
-  const [didCommandFillError, setDidCommandFillError] = useState(false);
-
   const [textfieldsValues, setTextfieldsValues] = useState(data.fields || {});
   const [fieldVisibility, setFieldVisibility] = useState(data.fields_visibility || {});
+
+  // Whether the text fields should be in a loading state
+  const [isLoading, setLoading] = useState(false);
 
   // Suggestions generated by the autofill system.
   const [previousBase, setPreviousBase] = useState([]);
@@ -62,57 +55,6 @@ const TextFieldsNode = ({ data, id }) => {
       return 'f0';
     }
   }, []);
-
-  const handleCommandFill = () => {
-    setIsCommandFillLoading(true);
-    setDidCommandFillError(false);
-    autofill(
-      Object.values(textfieldsValues),
-      commandFillNumber
-    ).then((new_fields) => {
-      // Add all the new fields to the existing ones
-      for (const field of new_fields) {
-        const buffer = textfieldsValues;
-        const uid = getUID(buffer);
-        buffer[uid] = field;
-        setTextfieldsValues(buffer);
-        setDataPropsForNode(id, { fields: buffer });
-        pingOutputNodes(id);
-      }
-    }).catch(e => {
-      if (e instanceof AIError) {
-        setDidCommandFillError(true);
-      } else {
-        throw new Error("Unexpected error: " + e);
-      }
-    }).finally(() => setIsCommandFillLoading(false));
-  };
-
-  const handleGenerateAndReplace = () => {
-    setDidGenerateAndReplaceError(false);
-    setGenerateAndReplaceIsLoading(true);
-    generateAndReplace(
-      generateAndReplacePrompt,
-      generateAndReplaceNumber,
-      generateAndReplaceIsUnconventional
-    ).then((fields) => {
-      const buffer = {};
-      for (const field of fields) {
-        const uid = getUID(buffer);
-        buffer[uid] = field;
-        setTextfieldsValues(buffer);
-        setDataPropsForNode(id, { fields: buffer });
-        pingOutputNodes(id);
-      }
-    }).catch(e => {
-      if (e instanceof AIError) {
-        console.log(e);
-        setDidGenerateAndReplaceError(true);
-      } else {
-        throw new Error("Unexpected error: " + e);
-      }
-    }).finally(() => setGenerateAndReplaceIsLoading(false));
-  }
 
   // Handle delete text field.
   const handleDelete = useCallback((event) => {
@@ -317,59 +259,42 @@ const TextFieldsNode = ({ data, id }) => {
     setPlaceholder(suggestedRows[0]);
   }, [suggestedRows, isSuggestionsLoading]);
 
-  const extendUI = (
-    <Stack gap={1}>
-      {didCommandFillError ?
-        <Text size="xs" c="red">
-          Failed to generate. Please try again.
-        </Text>
-        : <></>}
-      <NumberInput label="Rows" min={1} max={10} defaultValue={3} value={commandFillNumber} onChange={setCommandFillNumber}/>
-      <Button size="sm" variant="light" color="grape" fullWidth onClick={handleCommandFill} disabled={!enoughRowsForSuggestions()} loading={isCommandFillLoading}>Extend</Button>
-      {enoughRowsForSuggestions() ? <></>
-      : <Text size="xs" c="grape">
-          Enter at least 2 rows to generate suggestions.
-        </Text>}
-    </Stack>
-  )
+  // Add the entire list of `fields` to `textfieldsValues`
+  function addMultipleFields(fields) {
+    for (const field of fields) {
+      const buffer = textfieldsValues;
+      const uid = getUID(buffer);
+      buffer[uid] = field;
+      setTextfieldsValues(buffer);
+      setDataPropsForNode(id, { fields: buffer });
+      pingOutputNodes(id);
+    }
+  }
 
-  const replaceUI = (
-    <Stack gap={1}>
-      {didGenerateAndReplaceError ?
-        <Text size="xs" c="red">
-          Failed to generate. Please try again.
-        </Text>
-        : <></>}
-      <TextInput label="Generate a list of..." value={generateAndReplacePrompt} onChange={(e) => setGenerateAndReplacePrompt(e.currentTarget.value)}/>
-      <NumberInput label="Rows" min={1} max={10} defaultValue={3} value={generateAndReplaceNumber} onChange={setGenerateAndReplaceNumber}/>
-      <Switch color="grape" label="Make outputs unconventional" value={generateAndReplaceIsUnconventional} onChange={(e) => setGenerateAndReplaceIsUnconventional(e.currentTarget.checked)}/>
-      <Button size="sm" variant="light" color="grape" fullWidth onClick={handleGenerateAndReplace} loading={generateAndReplaceIsLoading}>Replace</Button>
-    </Stack>
-  )
+  // Replace the entirety of `textfieldValues` with `newFields`
+  function replaceFields(fields) {
+    const buffer = {};
+      for (const field of fields) {
+        const uid = getUID(buffer);
+        buffer[uid] = field;
+        setTextfieldsValues(buffer);
+        setDataPropsForNode(id, { fields: buffer });
+        pingOutputNodes(id);
+      }
+  }
 
-  const aiPopover = (
-    <Tabs color="grape" defaultValue="extend">
-      <Tabs.List grow>
-        <Tabs.Tab value="extend">
-          Extend
-        </Tabs.Tab>
-        <Tabs.Tab value="replace">
-          Replace
-        </Tabs.Tab>
-      </Tabs.List>
-      <Tabs.Panel value="extend" pb="xs">
-        {extendUI}
-      </Tabs.Panel>
-      <Tabs.Panel value="replace" pb="xs">
-        {replaceUI}
-      </Tabs.Panel>
-    </Tabs>
-  )
+  const aiPopover = <AiPopover
+    values={textfieldsValues}
+    addValues={addMultipleFields}
+    replaceValues={replaceFields}
+    areValuesLoading={isLoading}
+    setValuesLoading={setLoading}
+  />
 
   return (
     <BaseNode classNames="text-fields-node" nodeId={id}>
       <NodeLabel title={data.title || 'TextFields Node'} nodeId={id} icon={<IconTextPlus size="16px" />} aiPopoverContent={aiPopover} />
-      <Skeleton visible={generateAndReplaceIsLoading}>
+      <Skeleton visible={isLoading}>
         <div ref={setRef}>
           {Object.keys(textfieldsValues).map(i => (
             <div className="input-field" key={i}>
